+++ src/lib.rs
@@ -1,15 +1,14 @@
 //#[deny(missing_docs)]
 
-pub mod app;
+mod app;
+
 pub mod context;
 pub mod middleware;
 pub mod router;
 
-pub use crate::app::App;
-pub use crate::context::Context;
-pub use crate::middleware::Middleware;
-pub use http::{response::Builder, StatusCode};
-pub use hyper::{header, Body};
+pub use app::App;
+pub use http::response::Builder;
+pub use hyper::{header, Body, Request, Response, StatusCode};
 
 #[cfg(test)]
 mod tests {}
